backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
python-docx==1.1.2
python-multipart==0.0.12
httpx==0.27.2
openai==1.51.0
anthropic==0.36.1
python-dotenv==1.0.1
aiofiles==24.1.0
pypdf>=4.0.0
